requirements-vision-gateway.txt
fastapi==0.134.0
uvicorn==0.38.0
pydantic>=2.7.4,<3.0.0
certifi>=2025.0.0
google-genai>=1.30.0,<2.0.0
